pyproject.toml
[tool.poetry]
name = "prefect_docker"
version = "3.1.3"
description = "Workflow for metadata ingestion into Dataverse"
authors = ["Fjodor van Rijsselberg"]

[tool.poetry.dependencies]
python = "^3.11.0"
prefect = "^3.0.10"
python-dotenv = "^1.0.1"
boto3 = "^1.35.44"
dynaconf = "^3.2.6"
jmespath = "^1.0.1"
pydantic = "^2.9.2"
pydataverse = "^0.3.4"
factory-boy = "^2.0.0"
Faker = "^37.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
